7307 Fix category urls in index for all scopes but the first
All category urls were using the base url of the first store which
was being indexed at a time. The problem was that the URL class was
being reused for all categories, and the scope was defined on first
call only. Now we are setting the scope explicitly for every category.