7447 Solve problems with indexing on some stores
Indexing aborted if the products which were assigned to the affected store
are not in the first chunk of products during indexing, because the first
product collection in \IntegerNet\Solr\Model\Bridge\PagedProductIterator
was empty and the module took that as a reason to abort processing. We
are now restricting the product IDs which we want to index to the active
store so now empty collections occur.